refactor(task models): move TaskStatus to models/Task and remove Workflo
Agent-Id: agent-c23f0a7a-e6af-49ea-865f-66be793b178d
Linked-Note-Id: 47b7ef44-dce5-4b3f-afd3-ccdeae462de5
