+++ tests/03-interdependent-tasks/workflow-factory.unit.test.ts
@@ -4,11 +4,10 @@ import { describe, it, expect, beforeEach, afterEach } from "vitest";
 import { DataSource } from "typeorm";
 import { Task } from "../../src/models/Task";
 import { Result } from "../../src/models/Result";
-import { Workflow } from "../../src/models/Workflow";
-import { TaskStatus } from "../../src/workers/taskRunner";
+import { Workflow, WorkflowStatus } from "../../src/models/Workflow";
+import { TaskStatus } from "../../src/models/Task";
 import {
   WorkflowFactory,
-  WorkflowStatus,
   WorkflowValidationError,
 } from "../../src/workflows/WorkflowFactory";
 import { ApiErrorCode } from "../../src/utils/errorResponse";
