+++ tests/06-workflow-results/lazy-patch-final-result.unit.test.ts
@@ -4,7 +4,7 @@ import { DataSource } from "typeorm";
 import { Task } from "../../src/models/Task";
 import { Result } from "../../src/models/Result";
 import { Workflow, WorkflowStatus } from "../../src/models/Workflow";
-import { TaskStatus } from "../../src/workers/taskRunner";
+import { TaskStatus } from "../../src/models/Task";
 import { applyLazyFinalResultPatch } from "../../src/routes/workflowRoutes";
 
 const buildDataSource = (): DataSource =>
